feat(landing, profile): profile on same page as lmanding, small style edits
