--- app/page.css
@@ -1,33 +1,0 @@
-h3 {
-    font-weight: 550;
-}
-
-h1 {
-  font-size: 8rem;
-} 
-
-a {
-  font-size: unset;
-}
-.profile-link {
-  color: var(--accent1);
-}
-.skills-link {
-  color: var(--accent2);
-}
-.projects-link {
-  color: var(--accent3);
-}
-
-@media screen and (max-width: 600px) {
-  h1 {
-    font-size: 6rem;
-  }
-  main > div {
-    align-items: center;
-  }
-  .button-wrapper {
-    margin-top: auto;
-    margin-bottom: 2rem;
-  }
-}
