--- app/profile/components/cv.tsx
@@ -1,5 +1,0 @@
-const CV = () => <div>
-    <h2>Simon MENARD - Mon CV</h2>
-</div>
-
-export default CV;
